requirements.txt
mysql-connector-python>=8.2.0
matplotlib>=3.8.0
pandas>=2.2.0
numpy>=1.26.0
tkcalendar>=1.6.0
Pillow>=10.0.0